moved statistics
