+++ src/kpi/statistics/issues/issue.service.ts
@@ -1,13 +1,13 @@
 import { Injectable, Logger } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
-import { RepositoryNameDto } from 'src/github-api/model/Repository';
-import { RepositoryDocument } from './schemas/repository.schema';
-import { msToDateString } from './statistics/dateUtil';
+import { RepositoryIdentifier } from 'src/repositories/model';
+import { RepositoryDocument } from 'src/repositories/model/schemas';
+import { msToDateString } from '../developerFocus/lib';
 
 @Injectable()
-export class StatisticService {
-  private readonly logger = new Logger(StatisticService.name);
+export class IssueService {
+  private readonly logger = new Logger(IssueService.name);
 
   constructor(
     @InjectModel('Repository')
@@ -15,207 +15,9 @@ export class StatisticService {
   ) {}
 
   /**
-   *
-   * @param repoIdent
-   * @param limit a maximum of 100 files is returned
-   */
-  async getMostChangedFiles(repoIdent: RepositoryNameDto, userLimit?: number) {
-    const limit = userLimit ? userLimit : 100;
-    this.logger.log(
-      `getting the ${limit} most changed files for ${repoIdent.owner}/${repoIdent.repo}`,
-    );
-    const filter = {
-      repo: repoIdent.repo,
-      owner: repoIdent.owner,
-    };
-
-    const group = {
-      _id: '$pullFiles.filename',
-      count: { $sum: 1 },
-    };
-
-    const getDiffs = {
-      from: 'diffs',
-      localField: 'diffs',
-      foreignField: '_id',
-      as: 'expandedDiffs',
-    };
-
-    const getPullFiles = {
-      from: 'pullrequestfiles',
-      localField: 'expandedDiffs.pullRequestFiles',
-      foreignField: '_id',
-      as: 'pullFiles',
-    };
-
-    const res: { _id: string; count: number }[] = await this.repoModel
-      .aggregate()
-      .match(filter)
-      .unwind('$diffs')
-      .lookup(getDiffs)
-      .lookup(getPullFiles)
-      .unwind('$pullFiles')
-      .group(group)
-      .sort({ count: -1 }) //sort reverted
-      .limit(limit)
-      .exec();
-
-    let avg = 0;
-    res.forEach((e) => {
-      avg += e.count;
-    });
-    avg = avg / res.length;
-    this.logger.log(
-      `Calculation of most changed files for ${repoIdent.owner}/${repoIdent.repo} finished. Retrieved ${res.length} files. Average changes to the first files: ${avg}`,
-    );
-
-    return avg;
-  }
-
-  /**
-   * This method gives the count of the filenames that are changed together
-   * E.g.
-   * 1. pullRequestFiles: [File A, File B, File C, File D]
-   * 2. pullRequestFiles: [File A, File B, File E, File F]
-   * 3. pullRequestFiles: [File A, File B, File X, File Y]
-   * 4. pullRequestFiles: [File C, File D, File Z]
-   * output: Files A & B changed together 3 times
-   * @param repoIdent
-   * @param userLimit
-   */
-  async getFilesChangedTogether(repoIdent: RepositoryNameDto) {
-    const filter = {
-      repo: repoIdent.repo,
-      owner: repoIdent.owner,
-    };
-
-    const getDiffs = {
-      from: 'diffs',
-      localField: 'diffs',
-      foreignField: '_id',
-      as: 'expandedDiffs',
-    };
-
-    const getPullFiles = {
-      from: 'pullrequestfiles',
-      localField: 'expandedDiffs.pullRequestFiles',
-      foreignField: '_id',
-      as: 'pullFiles',
-    };
-
-    // TODO: we need to do this for all files
-    // most likely it would be a good idea to iterate
-    // all diffs once, and create a map/counter of which
-    // files occured together
-    //enter file name here
-    const file1 = ['package.json'];
-    const file2 = ['package-lock.json'];
-    const getFilesNames = {
-      $and: [
-        { 'pullFiles.filename': { $in: file1 } },
-        { 'pullFiles.filename': { $in: file2 } },
-      ],
-    };
-
-    const res: { _id: string; count: number }[] = await this.repoModel
-      .aggregate()
-      .match(filter)
-      .unwind('$diffs')
-      .lookup(getDiffs)
-      .lookup(getPullFiles)
-      .match(getFilesNames)
-      .exec();
-
-    this.logger.log(
-      `The files ${file1} & ${file2} are repeatedly changed together ${res.length} times.`,
-    );
-    return res.length;
-  }
-
-  /**
-   * Calculate the change in the pullrequests
-   * @param repoIdent
-   */
-  async sizeOfPullRequest(repoIdent: RepositoryNameDto) {
-    const filter = {
-      repo: repoIdent.repo,
-      owner: repoIdent.owner,
-    };
-
-    const getPullRequest = {
-      from: 'pullrequests',
-      localField: 'expandedDiffs.pullRequest',
-      foreignField: '_id',
-      as: 'expandedPullRequest',
-    };
-
-    const getDiffs = {
-      from: 'diffs',
-      localField: 'diffs',
-      foreignField: '_id',
-      as: 'expandedDiffs',
-    };
-
-    // we query the ids of the changed files (this is enough, because we just want to count the number)
-    // as well as the pull request number to sort the files and label them in the visualization
-    const res: {
-      _id: string;
-      changedFiles: string[];
-      pullRequestNumber: number;
-    }[] = await this.repoModel
-      .aggregate()
-      .match(filter)
-      .unwind('$diffs')
-      .lookup(getDiffs)
-      .lookup(getPullRequest)
-      .unwind('$expandedDiffs')
-      .unwind('$expandedPullRequest')
-      .project({
-        changedFiles: '$expandedDiffs.pullRequestFiles',
-        pullRequestNumber: '$expandedPullRequest.number',
-      })
-      .sort({ pullRequestNumber: 1 })
-      .exec();
-
-    const numberOfFiles = [];
-
-    res.forEach((e) => {
-      // somehow there exist pull requests without changed files. need to investigate
-      if ('changedFiles' in e) {
-        numberOfFiles.push(e.changedFiles.length);
-      }
-    });
-
-    const numberOfElements = numberOfFiles.length;
-    const avg =
-      numberOfFiles.reduce((acc, curr) => {
-        return acc + curr;
-      }, 0) / numberOfElements;
-
-    const variance = numberOfFiles.reduce((acc, curr) => {
-      return acc + Math.pow(curr - avg, 2) / numberOfElements;
-    }, 0);
-
-    const standardDeviation = Math.sqrt(variance);
-    this.logger.log(
-      `constiance ${variance} standard deviation ${standardDeviation}`,
-    );
-    this.logger.log(
-      `In average ${avg} files are changed with each pull request`,
-    );
-
-    return {
-      numberOfFiles,
-      avg,
-      variance,
-      standardDeviation,
-    };
-  }
-
-  /**
    * Number of issues with no assignees
    */
-  async numberOfAssignee(repoIdent: RepositoryNameDto) {
+  async numberOfAssignee(repoIdent: RepositoryIdentifier) {
     const filter = {
       repo: repoIdent.repo,
       owner: repoIdent.owner,
@@ -263,7 +65,7 @@ export class StatisticService {
    * Calculate the number of open tickets
    * @param repoIdent
    */
-  async numberOfOpenTickets(repoIdent: RepositoryNameDto) {
+  async numberOfOpenTickets(repoIdent: RepositoryIdentifier) {
     const filter = {
       repo: repoIdent.repo,
       owner: repoIdent.owner,
@@ -303,7 +105,7 @@ export class StatisticService {
    * find the tickets which are closed, if assignees is null count them, if assignees is not null count number of assignees
    * @param repoIdent
    */
-  async avgNumberOfAssigneeUntilTicketCloses(repoIdent: RepositoryNameDto) {
+  async avgNumberOfAssigneeUntilTicketCloses(repoIdent: RepositoryIdentifier) {
     const filter = {
       repo: repoIdent.repo,
       owner: repoIdent.owner,
@@ -376,7 +178,7 @@ export class StatisticService {
    * @param repoIdent
    */
   async avgTimeTillTicketWasAssigned(
-    repoIdent: RepositoryNameDto,
+    repoIdent: RepositoryIdentifier,
     userLimit?: number,
   ) {
     const limit = userLimit ? userLimit : 100;
@@ -448,7 +250,7 @@ export class StatisticService {
    * @param repoIdent
    * @param userLimit
    */
-  async workInProgress(repoIdent: RepositoryNameDto, userLimit?: number) {
+  async workInProgress(repoIdent: RepositoryIdentifier, userLimit?: number) {
     const limit = userLimit ? userLimit : 100;
 
     const filter = {
