+++ src/kpi/statistics/lib/repoQuery.ts
@@ -1,0 +1,8 @@
+import { RepositoryIdentifier } from 'src/repositories/model';
+
+export function getRepoFilter(repo: RepositoryIdentifier) {
+  return {
+    repo: repo.repo,
+    owner: repo.owner,
+  };
+}
