+++ src/kpi/statistics/lib/transformMapToObject.ts
